Remove about dialog (there's a sidebar in settings), reorder settings items, taller overlay mode
