+++ FluentScreenRecorder/FluentScreenRecorder.csproj
@@ -155,9 +155,6 @@
     </Compile>
     <Compile Include="Converters\SettingsConverters.cs" />
     <Compile Include="CustomMediaTransportControls.cs" />
-    <Compile Include="Dialogs\AboutDialog.xaml.cs">
-      <DependentUpon>AboutDialog.xaml</DependentUpon>
-    </Compile>
     <Compile Include="Dialogs\ChangelogDialog.xaml.cs">
       <DependentUpon>ChangelogDialog.xaml</DependentUpon>
     </Compile>
@@ -271,10 +268,6 @@
     <Page Include="Controls\SettingsDisplayControl.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
-    </Page>
-    <Page Include="Dialogs\AboutDialog.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="Dialogs\ChangelogDialog.xaml">
       <SubType>Designer</SubType>
